separated marker clusters from map component
