+++ src/styles/style.css
@@ -54,6 +54,7 @@ body a {
       grid-template-columns: repeat(auto-fit, minmax(500px, 1fr));
   margin: 0;
   border-top: 1px solid #ebebeb;
+  float: right;
 }
 
 .product-list {
