+++ src/styles/style.scss
@@ -30,6 +30,7 @@ body {
 .container {
   @include gridRepeat(500px);
   border-top: 1px solid #ebebeb;
+  float: right;
 }
 
 .product-list {
